Really closes: #7540.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@25896 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ uvm/hier/usr/lib/python2.5/reports/engine.py
@@ -407,12 +407,13 @@ def fix_hierarchy(output_base):
                         newDir = '%s/%s' % (one_day_dir, node_dir)
                         os.rename(node_path, newDir)
                         xmlFile = os.path.join(newDir, "report.xml")
-                        f = open(xmlFile)
-                        xml = f.read()
-                        f.close()
-                        f = open(xmlFile, 'w')
-                        f.write(re.sub(r'/(\d\d\d\d-\d\d-\d\d)/', r'/\1/1-day/', xml))
-                        f.close()
+                        if os.path.isfile(xmlFile):
+                            f = open(xmlFile)
+                            xml = f.read()
+                            f.close()
+                            f = open(xmlFile, 'w')
+                            f.write(re.sub(r'/(\d\d\d\d-\d\d-\d\d)/', r'/\1/1-day/', xml))
+                            f.close()
 
 def get_node_base(name, date_base, report_days=1, host=None, user=None,
                   email=None):
